Revert "Revert "Fix ABI source code position for ABI feature (#260)""
This reverts commit 2bd5d777d2c50e9bc0cfb1d62864086cd1417fb3.
