+++ src/main/java/com/algorand/algosdk/abi/TypeAddress.java
@@ -1,4 +1,4 @@
-package com.algorand.algosdk.util.abi;
+package com.algorand.algosdk.abi;
 
 import com.algorand.algosdk.crypto.Address;
 import org.apache.commons.lang3.ArrayUtils;
