+++ src/test/java/com/algorand/algosdk/abi/TestRandomValues.java
@@ -1,4 +1,4 @@
-package com.algorand.algosdk.util.abi;
+package com.algorand.algosdk.abi;
 
 import com.algorand.algosdk.crypto.Address;
 import com.algorand.algosdk.util.Encoder;
